Amed

+++ AI/midj_logic.py
@@ -1,4 +1,5 @@
 from aiogram import Bot, Dispatcher, types
+from aiogram.dispatcher.filters import CommandStart
 from aiogram.utils import executor
 from aiogram.types import InlineKeyboardMarkup, InlineKeyboardButton
 from aiogram.dispatcher.filters.state import State, StatesGroup
@@ -10,6 +11,8 @@ import asyncio
 import aiohttp
 import config
 from config import bot, dp
+import data.limits_manage as lm
+from aiogram.dispatcher.filters import Command, Text
 
 
 class Form(StatesGroup):
@@ -23,7 +26,12 @@ class Form(StatesGroup):
 MID_TOKEN = config.MID_TOKEN
 
 
-async def generate_image(prompt: str):
+async def generate_image(prompt: str, user_id):
+    user = lm.db_manager.get_user(user_id)  # Получаем информацию о пользователе
+    if not (user['mj_subscription_type'] and user['mj52_limit'] > 0):
+        # Возвращаем ошибку или сообщение о недостатке лимитов
+        return False, "У вас нет активной подписки Midjourney или закончились лимиты."
+
     url = "https://api.mymidjourney.ai/api/v1/midjourney/imagine"
     payload = {"prompt": prompt}
     headers = {
@@ -36,7 +44,9 @@ async def generate_image(prompt: str):
             await asyncio.sleep(15)
             if response.status == 200:
                 response_data = await response.json()
-                print(response_data)
+                lm.update_gpt_limits(user_id, 'midjourney5.2', 1)
+                print(response_data, 'лимит отнялся, миджорни')
+
                 await asyncio.sleep(15)
 
                 return await midle_check(response_data)
@@ -103,7 +113,7 @@ async def midle_check(response_data):
                 return False, False
 
 
-async def blend(images):
+async def blend(images, user_id):
     api_url = "https://api.mymidjourney.ai/api/v1/midjourney/blend"
     headers = {
         "Content-Type": "application/json",
@@ -114,7 +124,9 @@ async def blend(images):
         async with session.post(api_url, headers=headers, data=json.dumps(data)) as response:
             if response.status == 200:
                 response_data = await response.json()
-                print(response_data)
+                lm.update_gpt_limits(user_id, 'midjourney5.2', 1)
+
+                print(response_data, 'лимит отнялся, миджорни')
                 await asyncio.sleep(30)
                 return await midle_check(response_data)
             else:
@@ -148,8 +160,13 @@ async def block(input, state: FSMContext):
 
 @dp.message_handler(commands=['imagine'], state='*')
 async def process_start_command(message: types.Message, state: FSMContext):
-    await state.set_state(Form.waiting_for_user_input)
-    await message.reply("Введи промт")
+    user_id = message.from_user.id
+    user = lm.db_manager.get_user(user_id)  # Предполагаем, что функция get_user возвращает информацию о пользователе
+    if user['mj_subscription_type'] and user['mj52_limit'] > 0:
+        await state.set_state(Form.waiting_for_user_input)
+        await message.reply("Введи промт")
+    else:
+        await message.reply("У вас нет активной подписки Midjourney или закончились лимиты.")
 
 
 @dp.message_handler(commands=['face_swap'], state='*')
@@ -161,22 +178,33 @@ async def face_swap_main(message: types.Message, state: FSMContext):
 
 @dp.message_handler(commands=['blend'], state='*')
 async def start_blend(message: types.Message, state: FSMContext):
-    # user_id = message.from_user.id
-    # processing_tasks.pop(user_id, None) #для блока с фото
-    # photos_cash.pop(user_id, None)
+    user_id = message.from_user.id
+    user = lm.db_manager.get_user(user_id)
+    if user['mj_subscription_type'] and user['mj52_limit'] > 0:
 
-    await state.finish()
-    await state.set_state(Form.blend)
-    await message.reply("Отправьте до 5 ссылок на изображения, разделённых запятыми.")
+        # user_id = message.from_user.id
+        # processing_tasks.pop(user_id, None) #для блока с фото
+        # photos_cash.pop(user_id, None)
+        await state.finish()
+        await state.set_state(Form.blend)
+        await message.reply("Отправьте до 5 ссылок на изображения, разделённых запятыми.")
+    else:
+        await message.reply("У вас нет активной подписки Midjourney или закончились лимиты.")
+        return
 
 
 @dp.message_handler(state=Form.waiting_for_user_input)
 async def user_input(message: types.Message, state: FSMContext, image_url=None, message_id=None, time_message=None):
+    if message.text.startswith('/'):
+        await state.finish()
+        return
+
+    user_id = message.from_user.id
     user_text = message.text
     await state.set_state(Form.block)
     if image_url is None:
         time_message = await message.answer("Идёт генерация...")
-        image_url, message_id = await generate_image(user_text)
+        image_url, message_id = await generate_image(user_text, user_id)
         await time_message.delete()
 
     else:
@@ -209,13 +237,26 @@ async def process_callback(callback, state: FSMContext):
     message_id = info['message_id']
     await state.set_state(Form.block)
 
+    user_id = callback.from_user.id
+    callback_data = callback.data
+    user = lm.db_manager.get_user(user_id)
+
     time_message = await callback.message.answer("Идёт генерация...")
     image, message_id = await button_click(callback.data, message_id)
 
     if image:
+        if user['mj_subscription_type'] is None and not user['mj52_limit'] > 0:
+            await callback.message.answer("У вас нет активной подписки Midjourney или закончились лимиты.")
+            return
+
         inline_kb_full = InlineKeyboardMarkup()
         if callback.data in ['V1', 'V2', 'V3', 'V4', "🔄", 'Zoom Out 2x', 'Zoom Out 1.5x', 'Vary (Strong)',
                              'Vary (Subtle)']:
+
+            '''отнимаю лимит'''
+            lm.update_gpt_limits(user_id, 'midjourney5.2', 1)  # Обновляем лимиты для пользователя
+            print('лимит отнялся, миджорни')
+
             inline_kb_full.row(InlineKeyboardButton('U1', callback_data='U1'),
                                InlineKeyboardButton('U2', callback_data='U2'),
                                InlineKeyboardButton('U3', callback_data='U3'),
@@ -229,8 +270,15 @@ async def process_callback(callback, state: FSMContext):
             if callback.data in ['Upscale (4x)', 'Upscale (2x)', 'Redo Upscale (2x)', 'Redo Upscale (4x)']:
                 inline_kb_full.row(InlineKeyboardButton('Redo Upscale (2x)', callback_data="Redo Upscale (2x)"),
                                    InlineKeyboardButton("Redo Upscale (4x)", callback_data="Redo Upscale (4x)"))
+                '''отнимаю лимит'''
+                lm.update_gpt_limits(user_id, 'midjourney5.2', 1)  # Обновляем лимиты для пользователя
+                print('лимит отнялся, миджорни')
 
             elif callback.data in ['⬅️', '➡️', '⬆️', '⬇️']:
+                '''отнимаю лимит'''
+                lm.update_gpt_limits(user_id, 'midjourney5.2', 1)  # Обновляем лимиты для пользователя
+                print('лимит отнялся, миджорни')
+
                 inline_kb_full.row(InlineKeyboardButton('U1', callback_data='U1'),
                                    InlineKeyboardButton('U2', callback_data='U2'),
                                    InlineKeyboardButton('U3', callback_data='U3'),
@@ -238,6 +286,10 @@ async def process_callback(callback, state: FSMContext):
                                    InlineKeyboardButton('🔄', callback_data='🔄'))
 
             else:
+                '''отнимаю лимит'''
+                lm.update_gpt_limits(user_id, 'midjourney5.2', 1)  # Обновляем лимиты для пользователя
+                print('лимит отнялся, миджорни')
+
                 inline_kb_full.row(InlineKeyboardButton('Upscale (2x)', callback_data="Upscale (2x)"),
                                    InlineKeyboardButton("Upscale (4x)", callback_data="Upscale (4x)"),
                                    InlineKeyboardButton("Zoom Out 2x", callback_data="Zoom Out 2x"),
@@ -261,7 +313,8 @@ async def process_callback(callback, state: FSMContext):
         await state.set_state(Form.waiting_user_btn)
 
 
-# ЭТО БЛОК ДЛЯ ФОТО АЛЬБОМОМ
+"""ЭТО БЛОК ДЛЯ ФОТО АЛЬБОМОМ"""
+
 
 # processing_tasks = {}  # Словарь для отслеживания задач обработки по пользователю
 # photos_cash = {}
@@ -293,6 +346,11 @@ async def process_callback(callback, state: FSMContext):
 
 @dp.message_handler(content_types=['text'], state=Form.blend)
 async def blend_get_links(message: types.Message, state: FSMContext):
+    if message.text.startswith('/'):
+        await message.reply("Это команда. Пожалуйста, отправьте ссылки на изображения.")
+        await state.finish()
+
+        return
     links = message.text.replace(' ', '').replace('\n', '').split(',')
     if len(links) > 5:
         await message.reply("Пожалуйста, отправьте не более 5 ссылок.")
@@ -304,7 +362,7 @@ async def blend_get_links(message: types.Message, state: FSMContext):
 async def process_images(message, state, images):
     await state.set_state(Form.block)
     time_message = await message.answer("Идёт генерация...")
-    response, messageId = await blend(images)
+    response, messageId = await blend(images, message.from_user.id)
     await user_input(message, state, response, messageId, time_message)
 
 
@@ -319,3 +377,12 @@ async def face_swap_get_links(message: types.Message, state: FSMContext):
         response, messageId = await face_swap(links[0], links[1])
         await user_input(message, state, response, messageId, time_message)
 
+
+@dp.message_handler(Text(startswith="/"), state="*")
+async def handle_any_command(message: types.Message, state: FSMContext):
+    current_state = await state.get_state()
+    if current_state is not None:
+        # Сброс состояния пользователя
+        await state.finish()
+        # Сообщение пользователю, что его предыдущее действие было отменено
+        await message.reply("Текущая операция была отменена. Выберите команду для продолжения.")
